lots of rework to extract common modules/utils.
