+++ lmplay/exp/weights/uw1/versions.py
@@ -1,0 +1,12 @@
+from lmplay.base.base_model import BasicModelRunner
+from .model import GPT2
+from lmplay.base.runner_list import expose_runner
+
+
+
+@expose_runner('uw1_0', description="Unifeid Weights using the basic mbias, mbias_bias and bias_bias")
+def runner(*args, **kwargs):
+  return BasicModelRunner(GPT2,
+                          *args,
+                          overrides=dict(),
+                          **kwargs)
