+++ lmplay/exp/weights/uw2/modules.py
@@ -10,9 +10,7 @@ def gen_mask(max_len: int) -> torch.Tensor:
   return torch.tril(torch.ones(max_len, max_len)).unsqueeze(0).unsqueeze(0)
 
 
-class SULinear(nn.Module):
-  # Modified from pytorch source
-  #This combines UW 1.0 and UW 2.0 so shared sacrificial bias gen and mbias
+class ULinear(nn.Module):
   def __init__(self,
                shared_mid_weights: nn.Linear,
                in_features: int,
@@ -24,13 +22,10 @@ class SULinear(nn.Module):
     self.in_features = in_features
     self.out_features = out_features
     self.weight = nn.Parameter(torch.empty((out_features, in_features), **factory_kwargs))
-    #Hack to avoid storing copies of the mid weights everywhere
+    #Hack to avoid saving copies of the shared mid weights everywhere
     self.shared_mid_weights = [shared_mid_weights]
     self.expansion_data = nn.Parameter(torch.empty(shared_mid_weights.in_features))
     self.bias_weights = nn.Linear(shared_mid_weights.out_features, out_features)
-    self.mbias = nn.Parameter(torch.ones(out_features, **factory_kwargs))
-    self.mbias_bias = nn.Parameter(torch.zeros(1, **factory_kwargs))
-    self.bias_bias = nn.Parameter(torch.zeros(1, **factory_kwargs))
     self.reset_parameters()
 
   def reset_parameters(self) -> None:
@@ -47,7 +42,7 @@ class SULinear(nn.Module):
     result = F.linear(input, self.weight, None)
     bias = self.shared_mid_weights[0](self.expansion_data)
     bias = self.bias_weights(F.gelu(bias))
-    result = result * (self.mbias + self.mbias_bias) + (bias + self.bias_bias)
+    result = result + bias
     return result
 
   def extra_repr(self) -> str:
