+++ lmplay/exp/weights/uw4/unified_weights_v4_1/model.py
@@ -2,20 +2,11 @@ import torch
 from torch import nn
 from typing import Optional, Any, List
 
-from lmplay.exp.weights.modules import ULinear, DULinear
-# from .modules import DULinear
-from lmplay.modules import Block
+from lmplay.modules import Block, DULinear
 import tiktoken
 from lmplay.base.base_model import LMBase, LMRunnerBase
 from functools import partial
-
-
-# See the ULinear in the modules for more info on how this works.
-
-def _p(v) -> str:
-  if v is None:
-    return 'N'
-  return int(v)
+#See the ULinear in the modules for more info on how this works.
 
 class GPT2(LMBase):
   def __init__(self,
@@ -26,32 +17,19 @@ class GPT2(LMBase):
                attn_dropout: Optional[float] = 0.1,
                ff_dropout: Optional[float] = 0.1,
                embed_dropout: Optional[float] = 0.1,
-               version="5.1",
-               exp_mul=6.0,
-               mid_mul=1.0,
-               predict_bias=True,
-               predict_mbias=True,
-               predict_mbias2=True,
-               predict_mbias_a=False,  #Without a better init for the weights we can't set this to True
-               predict_mbias2_a=False, #Without a better init for the weights we can't set this to True
-               ln_attn=False,
-               ln_mlp=False,
-               ln_fc=True,
-               dl_fc=True,
-               expansion_weights=True,
+               version="4.1",
                **ignore):
-    super().__init__(
-      f"uw_v{version}_{_p(predict_bias)}{_p(predict_mbias)}{_p(predict_mbias2)}{_p(predict_mbias_a)}{_p(predict_mbias2_a)}{_p(ln_attn)}{_p(ln_mlp)}{_p(ln_fc)}{_p(dl_fc)}{_p(expansion_weights)}_{exp_mul}_{mid_mul}_{num_blocks}L_{max_len}",
-      max_len=max_len,
-      num_heads=num_heads,
-      num_blocks=num_blocks,
-      embed_dim=embed_dim,
-      attn_dropout=attn_dropout,
-      ff_dropout=ff_dropout,
-      embed_dropout=embed_dropout,
-      version=version,
-      **ignore)
-    # Testing standard DULinear on just the blocks.
+    super().__init__(f"uw_v{version}_{num_blocks}L_{max_len}",
+                     max_len=max_len,
+                     num_heads=num_heads,
+                     num_blocks=num_blocks,
+                     embed_dim=embed_dim,
+                     attn_dropout=attn_dropout,
+                     ff_dropout=ff_dropout,
+                     embed_dropout=embed_dropout,
+                     version=version,
+                     **ignore)
+    #Trying predicting the mbias again. It didn't help before but with a exp_mul and separate sacrificial net it might.
     self.tokenizer = tiktoken.get_encoding("gpt2")
     vocab_size = self.tokenizer.n_vocab
 
@@ -59,33 +37,19 @@ class GPT2(LMBase):
     self.tok_embed = nn.Embedding(vocab_size, embed_dim)
     self.pos_embed = nn.Parameter(torch.zeros(1, max_len, embed_dim))
     self.dropout = nn.Dropout(embed_dropout)
-    # add in the DULinear to the block definition
+    #add in the DULinear to the block definition
     dulinear = partial(DULinear,
-                       exp_mul=exp_mul,
-                       mid_mul=mid_mul,
-                       predict_bias=predict_bias,
-                       predict_mbias=predict_mbias,
-                       predict_mbias2=predict_mbias2,
-                       predict_mbias_a=predict_mbias_a,
-                       predict_mbias2_a=predict_mbias2_a,
-                       expansion_weights=True,
-                       linear=nn.Linear)
+                       predict_mbias2=False,
+                       predict_bas2=False,
+                       predict_mbias=False)
     self.blocks = nn.Sequential(*[Block(max_len,
                                         num_heads,
                                         embed_dim,
                                         attn_dropout=attn_dropout,
                                         ff_dropout=ff_dropout,
-                                        linear=dulinear,
-                                        ln_attn=ln_attn,
-                                        ln_mlp=ln_mlp) for _ in range(num_blocks)])
-    if ln_fc:
-      self.ln = nn.LayerNorm(embed_dim)
-    else:
-      self.ln = lambda x:x
-    if dl_fc:
-      self.fc = dulinear(embed_dim, vocab_size)
-    else:
-      self.fc = nn.Linear(embed_dim, vocab_size)
+                                        linear=dulinear) for _ in range(num_blocks)])
+    self.ln = nn.LayerNorm(embed_dim)
+    self.fc = dulinear(embed_dim, vocab_size)
 
   def forward(self, x: torch.Tensor, cache: Optional[List] = None):
     seq_len = x.size(1)
