+++ lmplay/exp/weights/uw4/unified_weights_v4_2/model.py
@@ -2,11 +2,10 @@ import torch
 from torch import nn
 from typing import Optional, Any, List
 
-from .modules import ULinear
-from lmplay.modules import Block
+from lmplay.modules import Block, ULinear, DULinear
 import tiktoken
 from lmplay.base.base_model import LMBase, LMRunnerBase
-
+from functools import partial
 #See the ULinear in the modules for more info on how this works.
 
 class GPT2(LMBase):
@@ -18,8 +17,7 @@ class GPT2(LMBase):
                attn_dropout: Optional[float] = 0.1,
                ff_dropout: Optional[float] = 0.1,
                embed_dropout: Optional[float] = 0.1,
-               version="1.0",
-
+               version="4.2",
                **ignore):
     super().__init__(f"uw_v{version}_{num_blocks}L_{max_len}",
                      max_len=max_len,
@@ -31,7 +29,7 @@ class GPT2(LMBase):
                      embed_dropout=embed_dropout,
                      version=version,
                      **ignore)
-
+    #4.1 but with ULinear helping the sacrifical bias/mbias prediction
     self.tokenizer = tiktoken.get_encoding("gpt2")
     vocab_size = self.tokenizer.n_vocab
 
@@ -39,15 +37,19 @@ class GPT2(LMBase):
     self.tok_embed = nn.Embedding(vocab_size, embed_dim)
     self.pos_embed = nn.Parameter(torch.zeros(1, max_len, embed_dim))
     self.dropout = nn.Dropout(embed_dropout)
-    #add in the ULinear to the block definition
+    #add in the DULinear to the block definition
+    dulinear = partial(DULinear,
+                       predict_mbias2=False,
+                       predict_bias=False,
+                       linear=ULinear)
     self.blocks = nn.Sequential(*[Block(max_len,
                                         num_heads,
                                         embed_dim,
                                         attn_dropout=attn_dropout,
                                         ff_dropout=ff_dropout,
-                                        linear=ULinear) for _ in range(num_blocks)])
+                                        linear=dulinear) for _ in range(num_blocks)])
     self.ln = nn.LayerNorm(embed_dim)
-    self.fc = ULinear(embed_dim, vocab_size)
+    self.fc = dulinear(embed_dim, vocab_size)
 
   def forward(self, x: torch.Tensor, cache: Optional[List] = None):
     seq_len = x.size(1)
@@ -76,9 +78,6 @@ class GPT2(LMBase):
     return x
 
 
-from lmplay.base.runner_list import expose_runner
-
-@expose_runner('uw1_0', description="Unifeid Weights using the basic mbias, mbias_bias and bias_bias")
 class ModelRunner(LMRunnerBase):
   def __init__(self, max_batch_size=25):
     super().__init__(max_batch_size=max_batch_size)
