+++ lmplay/stats/plot.py
@@ -8,6 +8,7 @@ import multiprocessing as mp
 
 import numpy
 import numpy as np
+import bisect
 
 def find_stat_files(path: str) -> (tuple, tuple):
   path = os.path.expanduser(path)
@@ -33,81 +34,125 @@ def find_stat_files(path: str) -> (tuple, tuple):
 def _plot_worker(out_file,
                  file_data: dict,
                  plot_targets,
-                 log_plot:bool,
-                 show:bool,
+                 log_plot: bool,
+                 show: bool,
                  scale: bool,
-                 min_iter:int,
-                 max_iter:int,
-                 abs_min_value:float,
-                 abs_max_value:float,
-                 max_min_value:float,
-                 min_max_value:float,
-                 min_show:float,
-                 plot_raw:bool,
-                 average_count:Optional[int],
-                 satart_ac = 5,
-                 ac_inc = 1):
-  #with plt.xkcd():
+                 min_iter: int,
+                 max_iter: int,
+                 abs_min_value: float,
+                 abs_max_value: float,
+                 max_min_value: float,
+                 min_max_value: float,
+                 min_show: float,
+                 plot_raw: bool,
+                 average_count: Optional[int],
+                 satart_ac=5,
+                 ac_inc=1):
+  # with plt.xkcd():
   acs = dict()
+  #This is horrible nog good bad code. Sorry. Please don't think of this as something to use.
+  # It is something I tweak all the time to get things to look the way I want and not good code.
+  #def get_ac(i):
+  #  if i not in acs:
+  #    ac = []
+  #    center = int(i / 2)
+  #    for j in range(i * 2 + 1):
+  #      weight = (1.0 - abs(j - center) / i) ** 2
+  #      ac.append(weight)
+  #    ac = numpy.array(ac)
+  #    acs[i] = ac
+  #  return acs[i]
 
   def get_ac(i):
     if i not in acs:
       ac = []
-      center = int(i/2)
-      for j in range(i*2 + 1):
+      center = int(i / 2)
+      for j in range(i):
         weight = (1.0 - abs(j - center) / i) ** 2
         ac.append(weight)
       ac = numpy.array(ac)
       acs[i] = ac
     return acs[i]
-  #ac = ac/np.sum(ac)
+
+
+  # ac = ac/np.sum(ac)
   abs_avg_min_value = None
   abs_avg_max_value = None
+  abs_min_value = None
+  abs_max_value = None
   addtl_lines = []
   for name, data in file_data.items():
+    iters = data['iter']
+    first_d = 0
+    #bisect just gives a value which is stupid. I want an index
+    while first_d < len(iters) and iters[first_d] < min_iter:
+      first_d += 1
+    #first_d = 0
     for data_name in plot_targets:
       d = numpy.array(data[data_name])
-      #if average_count is None:
+      # if average_count is None:
       #  plt.plot(data['iter'], d, label=f"{name}_{data_name}", linewidth=1)
-      #else:
+      # else:
 
       if plot_raw:
         l = plt.plot(data['iter'], d, linewidth=.2, alpha=.3)
       else:
         l = None
-      avgs = [0.0]*len(d)
+      avgs = [0.0] * len(d)
       current_average_count = satart_ac
-      #inc = max(int(len(d)/average_count), 1)
-      inc = ac_inc*4
-      for i in range(len(d)):
-        if (i+1)%inc == 0:
-          current_average_count = min(average_count, current_average_count+ac_inc)
-        ac = get_ac(current_average_count)
-        d_start = max(0, i - current_average_count)
+      # inc = max(int(len(d)/average_count), 1)
+      inc = ac_inc * 4
+      for _ in range(int(first_d/inc)):
+        current_average_count = min(average_count, current_average_count + ac_inc)
+      for i in range(first_d, len(d)):
+        if (i + 1) % inc == 0:
+          current_average_count = min(average_count, current_average_count + ac_inc)
+        #ac = get_ac(current_average_count)
+        # Start and end are centered around i
         d_end = min(i + current_average_count, len(d))
-        count = min(i - d_start, d_end - i - 1)
-        d_start = i - count
-        d_end = i + count + 1
-        ac_start = current_average_count -count
-        ac_end = current_average_count + count + 1
+        d_start = max(0, d_end - current_average_count * 2)
+        d_end = min(len(d), d_start + current_average_count * 2)
+        count = d_end - d_start
+        ac = get_ac(count)
+        # count = min(i - d_start, d_end - i - 1)
+        # d_start = i - count
+        # d_end = i + count + 1
+        #ac_start = current_average_count - count
+        #ac_end = current_average_count + count + 1
         d_sect = d[d_start:d_end]
-        ac_sect = ac[ac_start:ac_end]
-        ac_sect = ac_sect/np.sum(ac_sect)
-        avg = d_sect*ac_sect
+        #ac_sect = ac[ac_start:ac_end]
+        ac_sect = ac
+        ac_sect = ac_sect / np.sum(ac_sect)
+        avg = d_sect * ac_sect
         avg = np.sum(avg)
         avgs[i] = avg
       if abs_avg_min_value is None:
-        abs_avg_min_value = min(avgs)
-        abs_avg_max_value = max(avgs)
+        abs_avg_min_value = min(avgs[first_d:])
+        abs_avg_max_value = max(avgs[first_d:])
+
+
       else:
-        abs_avg_min_value = min(min(avgs), abs_avg_min_value)
-        abs_avg_max_value = max(max(avgs), abs_avg_max_value)
+        abs_avg_min_value = min(min(avgs[first_d:]), abs_avg_min_value)
+        abs_avg_max_value = max(max(avgs[first_d:]), abs_avg_max_value)
+      if abs_min_value is None:
+        abs_min_value = float(d[first_d:].min())
+        abs_max_value = float(d[first_d:].max())
+      else:
+        abs_min_value = min(abs_min_value, float(d[first_d:].min()))
+        abs_max_value = max(abs_max_value, float(d[first_d:].max()))
+      if abs_max_value == abs_min_value:
+        abs_max_value = None
+        abs_min_value = None
+      if abs_avg_min_value == abs_avg_max_value:
+        abs_avg_max_value = None
+        abs_avg_min_value = None
       if l is None:
-        addtl_lines.append({'iter':data['iter'], 'data':avgs, "label": f"{name}_{data_name}"})
+        addtl_lines.append({'iter': data['iter'], 'data': avgs, "label": f"{name}_{data_name}"})
       else:
-        addtl_lines.append({'iter':data['iter'], 'data':avgs, 'color':l[0].get_color(), "label": f"{name}_{data_name}"})
+        addtl_lines.append(
+          {'iter': data['iter'], 'data': avgs, 'color': l[0].get_color(), "label": f"{name}_{data_name}"})
   for line_info in addtl_lines:
-    #do these last so the show on top of the other line data
+    # do these last so the show on top of the other line data
     if 'color' in line_info:
       plt.plot(line_info['iter'], line_info['data'], label=line_info['label'], linewidth=.5, color=line_info['color'])
     else:
@@ -119,20 +164,24 @@ def _plot_worker(out_file,
   if not plot_raw:
     abs_min_value = abs_avg_min_value
     abs_max_value = abs_avg_max_value
+
   # We want to scale it just the lowest part to zoom in on it so set the max to min_show above out max_min
   if scale:
-    #upper_show = (abs_max_value - max_min_value) * min_show
-    #upper_show += max_min_value
-    #y_min = abs_min_value - .001 * upper_show
-    #y_max = upper_show
+    # upper_show = (abs_max_value - max_min_value) * min_show
+    # upper_show += max_min_value
+    # y_min = abs_min_value - .001 * upper_show
+    # y_max = upper_show
     if abs_max_value == abs_min_value:
+      if abs_min_value is None:
+        abs_min_value = 0.0
+        abs_max_value = 0.0
       abs_max_value += .1
     y_min = abs_min_value
     y_max = abs_max_value
     plt.ylim(y_min, y_max)
-    #x_min = max(min_iter - max(min_iter * .2, 100), 1)
+    # x_min = max(min_iter - max(min_iter * .2, 100), 1)
     x_min = max(min_iter, 1)
-    plt.xlim(x_min, max_iter*1.1)
+    plt.xlim(x_min, max_iter * 1.1)
   # plt.autoscale(enable=True, axis='both', tight=True)
   plt.grid(True)
   plt.legend(bbox_to_anchor=(1.05, 1),
@@ -152,7 +201,8 @@ def _plot(*args):
   proc.start()
   proc.join()
 
-def get_stats(file_data:dict, plot_targets:tuple, shortest_iter:int):
+
+def get_stats(file_data: dict, plot_targets: tuple, shortest_iter: int):
   abs_min_value = None
   abs_max_value = None
   max_min_value = None
@@ -171,7 +221,6 @@ def get_stats(file_data:dict, plot_targets:tuple, shortest_iter:int):
           dataset_min = min(dataset_min, value)
           dataset_max = max(dataset_max, value)
 
-
     # now figure out our tracked max /min values across datasets
     if abs_max_value is None:
       abs_max_value = dataset_max
@@ -187,14 +236,14 @@ def get_stats(file_data:dict, plot_targets:tuple, shortest_iter:int):
     max_min_value = max(max_min_value, dataset_min)
     min_max_value = min(min_max_value, dataset_max)
 
-    #upper_show = (dataset_max - dataset_min) * min_show + dataset_min
+    # upper_show = (dataset_max - dataset_min) * min_show + dataset_min
 
   return abs_min_value, abs_max_value, max_min_value, min_max_value
 
 
-def unify_points(file_data:dict, iters:list, plot_targets:tuple) -> dict:
+def unify_points(file_data: dict, iters: list, plot_targets: tuple) -> dict:
   new_iters = []
-  result_file = {'iter':new_iters}
+  result_file = {'iter': new_iters}
   for pt in plot_targets:
     file_idx = 0
     file_iters = file_data['iter']
@@ -205,21 +254,22 @@ def unify_points(file_data:dict, iters:list, plot_targets:tuple) -> dict:
       if file_idx < len(file_iters):
         result_values.append(file_values[file_idx])
         new_iters.append(iter)
-        #since the iters passed in contain all possible iters we are either on this iter or ahead of it by some amount.
+        # since the iters passed in contain all possible iters we are either on this iter or ahead of it by some amount.
         if file_iters[file_idx] == iter:
-          #looks like we were on it. Advance to the next one
+          # looks like we were on it. Advance to the next one
           file_idx += 1
       else:
         break
   return result_file
 
-def _get_iters(files_data:dict) -> (list, str, str):
+
+def _get_iters(files_data: dict) -> (list, str, str):
   found_iters = set()
   longest = None
   longest_len = None
   shortest = None
   shortest_len = None
-  #Find all the 'iters' in every file so we can build values for all files.
+  # Find all the 'iters' in every file so we can build values for all files.
   for name, data in files_data.items():
     found_iters.update(data['iter'])
     if longest is None or longest_len < data['iter'][-1]:
@@ -231,12 +281,14 @@ def _get_iters(files_data:dict) -> (list, str, str):
   iters = list(found_iters)
   iters.sort()
   return dict(iters=iters, longest=longest, shortest=shortest)
-def _diff_to_target(files_data:dict, target:str, plot_targets:tuple) -> dict:
+
+
+def _diff_to_target(files_data: dict, target: str, plot_targets: tuple) -> dict:
   target = files_data[target]
   result_files = dict()
-  #Then we subtract the longest one's value from their values to normalize against it.
+  # Then we subtract the longest one's value from their values to normalize against it.
   for file_name, data in files_data.items():
-    result_file = {'iter':data['iter'].copy()}
+    result_file = {'iter': data['iter'].copy()}
     result_files[file_name] = result_file
     for pt in plot_targets:
       target_values = []
@@ -245,10 +297,11 @@ def _diff_to_target(files_data:dict, target:str, plot_targets:tuple) -> dict:
         if len(target[pt]) > i:
           target_values.append(data[pt][i] - target[pt][i])
         else:
-          #Looks like the target is shorter than this run. We should trim it and not display the rest.
+          # Looks like the target is shorter than this run. We should trim it and not display the rest.
           result_file['iter'] = result_file['iter'][:i]
           break
   return result_files
+
 
 def get_file_data(*files, plot_targets=('loss', 'accuracy')) -> (dict, dict):
   file_data = dict()
@@ -281,29 +334,30 @@ def get_file_data(*files, plot_targets=('loss', 'accuracy')) -> (dict, dict):
         file_data[name] = data
   return file_data, _get_iters(file_data)
 
+
 def plot(out_file,
-         file_data:(dict, dict),
+         file_data: (dict, dict),
          min_show=.1,
          log_plot=True,
          show=False,
          plot_targets=('loss',),
          scale=True,
-         average_count:Optional[int] = 10,
+         average_count: Optional[int] = 10,
          diff_to_target=False,
          use_process=True,
          target=None,
-         plot_raw = False):
+         plot_raw=False):
   out_file = os.path.expanduser(out_file)
   # We want to center the graph on the interesting areas so we need to track the overall min/max and the worst min value
   # across all datasets we are plotting. Then we will show from the absolute min to abve the worst min but below the abs max.
 
   file_data, file_meta = file_data
-  #iters, longest, shortest = _get_iters(file_data)
-  #now we have a spot for every iter. Let's get normalized value for every point by going through them all.
-  #First we make sure that they all have values for every iter
-  file_data = {name:unify_points(fd, file_meta['iters'], plot_targets) for name, fd in file_data.items()}
-  min_iter = int(file_data[file_meta['shortest']]['iter'][-1]*(1.0 - min_show))
-  #min_iter = iters[0]
+  # iters, longest, shortest = _get_iters(file_data)
+  # now we have a spot for every iter. Let's get normalized value for every point by going through them all.
+  # First we make sure that they all have values for every iter
+  file_data = {name: unify_points(fd, file_meta['iters'], plot_targets) for name, fd in file_data.items()}
+  min_iter = int(file_data[file_meta['shortest']]['iter'][-1] * (1.0 - min_show))
+  # min_iter = iters[0]
   max_iter = file_meta['iters'][-1]
   if diff_to_target:
     if not target is None:
